Initial implementation of StarTree index in Pinot.
This change adds a SegmentGenerator implementation that generates a
StarTree index, which contains partial aggregates (SUM) for dimension
combinations, such that no more than k (configurable) records must be
scanned to answer any given aggregation query (as opposed to the whole
segment).

In addition to this, the raw data is sorted according to partial
dimension combination. This can be used to speed up existing queries by
filtering the document ID set.

If a user defines "starTreeIndexSpecs" in the schema (currently only one
allowed), then StarTreeSegmentCreator implementation will be used. The
existing implementation will be used otherwise.

RB=555274
R=kgopalak,dpatel
A=kgopalak