Disabling generation of aggregation / dimension stats.
When there are many mappers spawned (e.g. for a very large / partitioned
input.paths), this causes the namespace quota to be reached too quickly, as one
aggregation and one dimension stat JSON file is written for every task.

We need to think about how to manage this better, but since it is not being
used currently other than for debugging, will remove it.

RB=554880

R=kgopalak,npawar
A=npawar