[PINOT-2131] Fixed a bunch of memory allocation issues that caused too much GC
Changed the usage of Set<Integer> to be an IntSet instead. Specifically, the one
in ScanBasedSingleValueDocIdSet was showing up in this particular data set.

Changed the usage in other places as well.

Changed the Trace functionality to not allocate a Long object unless tracing is enabled.
This was the original suspect for the issue.

RB=574528
BUG=PINOT-2131
G=pinot-dev-reviewers
R=kgopalak,jfim,dpatel,mshrivas
A=jfim