[PINOT-2413] Fixing segment refcnt logic to increment/decrement under lock.
Removed unused calls and fold reference count into OfflineTableDataManager.

Introduced a read-write lock in OfflineSegmentDataManager. The logic to get a segment
for processing acquired readLock when doing so, and gets a handle to the SegmentDataManager.
Changed the segment release logic to release the SegmentDataManager instead of the segment name
string.

Added a unit test (and a stress test) to ensure correctness with multiple threads

RB=631784
BUG=PINOT-2413
G=pinot-dev-reviewers
R=kgopalak,jfim,atumbde,mshrivas
A=jfim