[PINOT-2451] Changed the segment refresh to be message based instead of toggling states
The server is now capable of handling a helix message that can be used to trigger a segment refresh.
The controller has been changed to send out this message instead of the usual ONLINE-to-OFFLINE-to-ONLINE
state transition on a selective basis. New segment adds (and segment drops) happen as before.

The controller looks for a table configuration setting (under metadata/customConfigs) to decide
whether to use the messaging or state transition for segment refresh. The table configuration is as:

        "messsageBasedRefresh" : "true"

If this config is found, then a message is sent to the server to refresh the segment.

Also changed a log message in the server to print the request Id, to match up query logs on the broker.