[PINOT-3089] Modify QueryRunner to make it able to test the system QPS with multi clients
The old multiThreadedQueryRunner doesn't work because driver.postQuery(query) is a
blocking call, so the targetQps makes no sense in this case.
Also, because the call() never returns, the future.get() will stuck there forever.

The fix is use QueryRunner to start multiple clients and collect QPS and avg
response time for a query set.

RB=730027
G=pinot-dev-reviewers
R=jfim,mshrivas,kgopalak,vradhakr,atumbde,raringun,ssubrama
A=mshrivas