Remove most log output from Travis builds (#192)
Travis builds are capped to 10k lines of log output. This patch adds a
new log4j configuration that is only used during Travis builds which
removes most of the output so that we can read the maven logs, which
contain the list of tests that failed, if applicable. Verified that
the Travis logs don't exceed the 10k line limit.

Reviewer: Jackie