[PINOT-3329] Fix exception in realtime consumer hashcode (#275)
KafkaConsumerManager has a map whose key is KafkaHighLevelStreamProviderConfig. One of the fields of
KafkaHighLevelStreamProviderConfig is of type Schema, which hashes the FieldSpec object. FieldSpec's
hash expects that getDefaultNullValue() works right, but when the Schema has a BOOLEAN field, an exception
is thrown when computing this hash.

Fixed FieldSpec.toString() (used in the hashcode method) to handle exception if getDefaultNullValue() returns
an exception.