Make the padding character configurable and change it to the null character. (#286)
By default the padding character will continue to be '%' but it can be
changed to any other character. This will automatically support segments generated with the  current default
'%' and with no padding character specified in the metadata.
This fixes the case when the column value has a trailing percent, which results in wrong results.