[PINOT-3203] Controller code changes for Low-level consumer segments (#322)
Updated PinotLLCRealtimeSegmentManager to support closing an old segment in propertystore
and creating new ones in idealstate. Added tests for the new methods.

Added a tentative protocol file (actual protocol elements TBF) and an FSM implementation
for closing segments when multiple servers inform the controller about their offsets, and
the controller ensures that one of them commits the segment.

Added test cases to test the scenarios described in the design documents.