[PINOT-3411]: Fix bug in tenant update that drops tenants while down (#342)
scaling.

Updating the broker tenant config can drop additional broker instances
when down scaling. A serious side-effect of this is that when Helix is
requested to drop this instance, it does not remove its reference from
the BrokerResource in IdealState. This breaks all future calls to query
instances with a specific tag from Helix. While this PR only addresses
the bug in Pinot, the Helix issue would/should be investigated
separately.