Fix the number of replicas in idealstate so that we get multiple servers in CONSUMING state (#420)
Setting the REPLICAS value to 1 in idealstate causes helix to send a state transition to
only one replica to go into CONSUMING state. We now set it to the number of replicas in
the IDEALSTATE so that helix sends state transition to all replicas.

Helix should still honor the fact that there is only one replica for the high-level
consumers. Need to check with a combination of high and low-level consumers  with multiple
servers