Replace assertEquals by assertTrue (#616)
Replace assertEquals by assertTrue(value.equals()) for the Netty
integration test where large requests/responses are generated. Since
this test generates a 2 MiB request and on assertion failure both the
expected value and received value are printed, it causes the Travis
build to error out due to exceeding the 4 MB limit. With this, the
build will still fail but the log will be complete instead of being
filled with request_iiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiii...